Remove GameRunner
