+++ .gitignore
@@ -72,7 +72,7 @@ classes.jar
 .externalToolBuilders/
 
 # Locally stored "Eclipse launch configurations"
-*.launch
+#*.launch
 
 # CDT-specific
 .cproject
