+++ src/games/strategy/debug/DebugUtils.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Properties;
 
 import games.strategy.engine.ClientContext;
-import games.strategy.engine.framework.GameRunner2;
+import games.strategy.engine.framework.GameRunner;
 
 /**
  * Moved out of Console class, so that we don't need swing.
@@ -70,7 +70,7 @@ public class DebugUtils {
     buf.append("Free memory: ").append(runtime.freeMemory() / mb).append("\r\n");
     buf.append("Total memory: ").append(runtime.totalMemory() / mb).append("\r\n");
     buf.append("Max memory: ").append(runtime.maxMemory() / mb).append("\r\n");
-    final int currentMaxSetting = GameRunner2.getMaxMemoryFromSystemIniFileInMB(GameRunner2.getSystemIni());
+    final int currentMaxSetting = GameRunner.getMaxMemoryFromSystemIniFileInMB(GameRunner.getSystemIni());
     if (currentMaxSetting > 0) {
       buf.append("Max Memory user setting within 22% of: ").append(currentMaxSetting).append("\r\n");
     }
