+++ src/games/strategy/engine/framework/EngineVersionProperties.java
@@ -115,7 +115,7 @@ public class EngineVersionProperties {
     final HostConfiguration config = client.getHostConfiguration();
     config.setHost(url.getHost());
     // add the proxy
-    GameRunner2.addProxy(config);
+    GameRunner.addProxy(config);
     final GetMethod method = new GetMethod(url.getPath());
     // pretend to be ie
     method.setRequestHeader("User-Agent", "Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.0)");
