+++ src/games/strategy/engine/framework/startup/launcher/ServerLauncher.java
@@ -25,7 +25,7 @@ import games.strategy.debug.ClientLogger;
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.PlayerID;
 import games.strategy.engine.framework.GameDataManager;
-import games.strategy.engine.framework.GameRunner2;
+import games.strategy.engine.framework.GameRunner;
 import games.strategy.engine.framework.ServerGame;
 import games.strategy.engine.framework.headlessGameServer.HeadlessGameServer;
 import games.strategy.engine.framework.message.PlayerListing;
@@ -192,7 +192,7 @@ public class ServerLauncher extends AbstractLauncher {
       if (m_abortLaunch) {
         m_serverReady.countDownAll();
       }
-      if (!m_serverReady.await(GameRunner2.getServerStartGameSyncWaitTime(), TimeUnit.SECONDS)) {
+      if (!m_serverReady.await(GameRunner.getServerStartGameSyncWaitTime(), TimeUnit.SECONDS)) {
         System.out.println("Waiting for clients to be ready timed out!");
         m_abortLaunch = true;
       }
@@ -239,8 +239,8 @@ public class ServerLauncher extends AbstractLauncher {
             try {
               // we are already aborting the launch
               if (!m_abortLaunch) {
-                if (!m_errorLatch.await(GameRunner2.getServerObserverJoinWaitTime()
-                    + GameRunner2.ADDITIONAL_SERVER_ERROR_DISCONNECTION_WAIT_TIME, TimeUnit.SECONDS)) {
+                if (!m_errorLatch.await(GameRunner.getServerObserverJoinWaitTime()
+                    + GameRunner.ADDITIONAL_SERVER_ERROR_DISCONNECTION_WAIT_TIME, TimeUnit.SECONDS)) {
                   System.err.println("Waiting on error latch timed out!");
                 }
               }
