+++ src/games/strategy/engine/framework/ui/SaveGameFileChooser.java
@@ -6,7 +6,7 @@ import javax.swing.JFileChooser;
 import javax.swing.filechooser.FileFilter;
 
 import games.strategy.engine.ClientContext;
-import games.strategy.engine.framework.GameRunner2;
+import games.strategy.engine.framework.GameRunner;
 import games.strategy.engine.framework.headlessGameServer.HeadlessGameServer;
 
 public class SaveGameFileChooser extends JFileChooser {
@@ -23,8 +23,8 @@ public class SaveGameFileChooser extends JFileChooser {
 
   public static String getAutoSaveFileName() {
     if (HeadlessGameServer.headless()) {
-      final String saveSuffix = System.getProperty(GameRunner2.TRIPLEA_NAME_PROPERTY,
-          System.getProperty(GameRunner2.LOBBY_GAME_HOSTED_BY, ""));
+      final String saveSuffix = System.getProperty(GameRunner.TRIPLEA_NAME_PROPERTY,
+          System.getProperty(GameRunner.LOBBY_GAME_HOSTED_BY, ""));
       if (saveSuffix.length() > 0) {
         return saveSuffix + "_" + AUTOSAVE_FILE_NAME;
       }
@@ -34,8 +34,8 @@ public class SaveGameFileChooser extends JFileChooser {
 
   public static String getAutoSave2FileName() {
     if (HeadlessGameServer.headless()) {
-      final String saveSuffix = System.getProperty(GameRunner2.TRIPLEA_NAME_PROPERTY,
-          System.getProperty(GameRunner2.LOBBY_GAME_HOSTED_BY, ""));
+      final String saveSuffix = System.getProperty(GameRunner.TRIPLEA_NAME_PROPERTY,
+          System.getProperty(GameRunner.LOBBY_GAME_HOSTED_BY, ""));
       if (saveSuffix.length() > 0) {
         return saveSuffix + "_" + AUTOSAVE_2_FILE_NAME;
       }
@@ -45,8 +45,8 @@ public class SaveGameFileChooser extends JFileChooser {
 
   public static String getAutoSaveOddFileName() {
     if (HeadlessGameServer.headless()) {
-      final String saveSuffix = System.getProperty(GameRunner2.TRIPLEA_NAME_PROPERTY,
-          System.getProperty(GameRunner2.LOBBY_GAME_HOSTED_BY, ""));
+      final String saveSuffix = System.getProperty(GameRunner.TRIPLEA_NAME_PROPERTY,
+          System.getProperty(GameRunner.LOBBY_GAME_HOSTED_BY, ""));
       if (saveSuffix.length() > 0) {
         return saveSuffix + "_" + AUTOSAVE_ODD_ROUND_FILE_NAME;
       }
@@ -56,8 +56,8 @@ public class SaveGameFileChooser extends JFileChooser {
 
   public static String getAutoSaveEvenFileName() {
     if (HeadlessGameServer.headless()) {
-      final String saveSuffix = System.getProperty(GameRunner2.TRIPLEA_NAME_PROPERTY,
-          System.getProperty(GameRunner2.LOBBY_GAME_HOSTED_BY, ""));
+      final String saveSuffix = System.getProperty(GameRunner.TRIPLEA_NAME_PROPERTY,
+          System.getProperty(GameRunner.LOBBY_GAME_HOSTED_BY, ""));
       if (saveSuffix.length() > 0) {
         return saveSuffix + "_" + AUTOSAVE_EVEN_ROUND_FILE_NAME;
       }
