+++ src/games/strategy/engine/lobby/server/GameDescription.java
@@ -6,7 +6,7 @@ import java.io.ObjectInput;
 import java.io.ObjectOutput;
 import java.util.Date;
 
-import games.strategy.engine.framework.GameRunner2;
+import games.strategy.engine.framework.GameRunner;
 import games.strategy.engine.framework.headlessGameServer.HeadlessGameServer;
 import games.strategy.net.INode;
 import games.strategy.net.Node;
@@ -53,7 +53,7 @@ public class GameDescription implements Externalizable, Cloneable {
   private String m_engineVersion;
   private String m_gameVersion;
   private String m_botSupportEmail =
-      HeadlessGameServer.getInstance() != null ? System.getProperty(GameRunner2.LOBBY_GAME_SUPPORT_EMAIL, "") : "";
+      HeadlessGameServer.getInstance() != null ? System.getProperty(GameRunner.LOBBY_GAME_SUPPORT_EMAIL, "") : "";
 
   // if you add a field, add it to write/read object as well for Externalizable
   public GameDescription() {}
