+++ src/games/strategy/triplea/ui/menubar/ViewMenu.java
@@ -43,7 +43,7 @@ import games.strategy.engine.data.properties.ColorProperty;
 import games.strategy.engine.data.properties.IEditableProperty;
 import games.strategy.engine.data.properties.NumberProperty;
 import games.strategy.engine.data.properties.PropertiesUI;
-import games.strategy.engine.framework.GameRunner2;
+import games.strategy.engine.framework.GameRunner;
 import games.strategy.engine.framework.startup.ui.MainFrame;
 import games.strategy.triplea.image.MapImage;
 import games.strategy.triplea.image.TileImageFactory;
@@ -144,7 +144,7 @@ public class ViewMenu {
         if (selectedValue.equals(currentKey)) {
           return;
         }
-        GameRunner2.setDefaultLookAndFeel(lookAndFeels.get(selectedValue));
+        GameRunner.setDefaultLookAndFeel(lookAndFeels.get(selectedValue));
         EventThreadJOptionPane.showMessageDialog(frame, "The look and feel will update when you restart TripleA",
             new CountDownLatchHandler(true));
       }
