+++ src/tools/map/xml/creator/MapXmlCreator.java
@@ -65,7 +65,7 @@ import org.xml.sax.SAXException;
 import games.strategy.ui.SwingAction;
 import games.strategy.debug.ClientLogger;
 import games.strategy.engine.data.GameParser;
-import games.strategy.engine.framework.GameRunner2;
+import games.strategy.engine.framework.GameRunner;
 import games.strategy.engine.framework.startup.mc.GameSelectorModel;
 import tools.image.FileSave;
 
@@ -266,7 +266,7 @@ public class MapXmlCreator extends JFrame {
     MapXmlCreator.log(Level.INFO, "Starting MapXMLCreator");
 
     // handleCommandLineArgs(args);
-    GameRunner2.setupLookAndFeel();
+    GameRunner.setupLookAndFeel();
     start();
   }
